==> 1 - Wine Enthusiast - Processamento e Tratamento de Dados.ipynb <==
### Etapas iniciais: importação de pacotes, definição de funções, carregamento de dados

import pandas as pd

def values_missing(df, only_missing=True):
    '''Calcula e retorna o percentual e o número de registros nulos no dataframe df'''
    missing = pd.concat({'percent_missing': 100* df.isnull().sum() / len(df),
                         'rows_missing': df.isnull().sum()}, axis=1)
    if only_missing:
        missing = missing[missing.percent_missing > 0]
    return missing.sort_values(by='percent_missing', ascending=False)


winemag = pd.read_csv('datasets/winemag_reviews.csv')
winemag.info()

### Filtragem
- O dataset originalmente continha 149.693 registros, os quais cobriam diversos tipos/categorias, como vinhos brancos, espumantes e tintos. Como o trabalho restringe-se a vinhos tintos, nessa etapa foram filtrados apenas os vinhos dessa categoria resultando em  87.574 registros.<br>

winemag = winemag[ winemag.Category == 'Red' ]

winemag.head()

### Tratamento de dados duplicados
 - Nenhum registro inteiramente duplicado foi encontrado;
 - Quando analisado pela coluna **Title** foram obtidos 314 registros duplicados (0.38% dos registros); desses 303 estavam duplicados também pela coluna **Taster**, indicando qu nos outros 9 casos o vinho foi objetivo foi objeto de análise por mais de um degustador; 85 dos registros duplicados possuiam também o valor do **Rating** distinto.
 - Analisando os registros duplicados no dataframe e acessando as URLs descritas no campo **url** é possível perceber que muitos dos registros duplicados devem-se a novas análises feitas sobre o vinho, seja pelo mesmo degustador ou não; na maioria dos casos a nota atribuída ao vinho permaneceu a mesma.
 - Foi optado por manter no dataset apenas o registro correspondente a análise mais recente publicada sobre o vinho, por entender que a análise mais recente é a versão mais definitiva do parecer sobre o vinho; reduzindo o número de registros para 82.260.

# Nenhum registro inteiramente duplicado
winemag[ winemag.duplicated() ]

# 314 reviews de vinho duplicados quando 
winemag.duplicated(subset=['Title']).sum()

# Exibindo parte do dataset duplicado (em conjunto com os "originais") para melhor compreender a duplicidade
winemag[ winemag.duplicated(subset=['Title'], keep=False) ].sort_values(by='Title').head(50)

print('Registros duplicados com mesmo Title e Taster: {}'
      .format( winemag.duplicated(subset=['Title', 'Taster']).sum() ))
print('Registros duplicados com mesmo Title e Rating: {}'
      .format( winemag.duplicated(subset=['Title', 'Rating']).sum() ))


winemag = winemag.drop_duplicates(subset=['Title'], keep='first')

winemag.info()

### Correções e transformações nos dados
- Os campos **Alcohol**, **Bottle Size** e **Price** foram convertidos de *object* para respetivamente: *float64*, *int64* e *float64*.

# Alcohol : transformação em número real substituindo o padrão regex [% ] por ''; e a vírgula pelo ponto.
winemag['Alcohol'] = winemag['Alcohol'].str.replace('[% ]', '', regex=True).str.replace(',', '.').astype('float64')

# Bottle Size : transformação em número inteiro (em ml) substituindo o padrão regex [mlML ] por ''.
print('=== Bottle Size - Valores - antes ===\n', winemag['Bottle Size'].value_counts())

winemag['Bottle Size'] = winemag['Bottle Size']\
        .str.replace('[mlML ]', '', regex=True).astype('float64')\
        .apply(lambda size: size * 1000 if size <= 3 else size).astype('int64') # Transforma L em ml

print('\n=== Bottle Size - Valores - depois ===\n', winemag['Bottle Size'].value_counts())

# Price : transformação em número real substituindo o padrão regex [R$. ] por ''; e a vírgula pelo ponto.
winemag['Price'] = winemag['Price'].str.replace('[R$. ]', '', regex=True).str.replace(',', '.').astype('float64')

### Tratamento de dados omissos
- Foram encontrados valores omissos em 6 colunas e dados os seguintes tratamentos:
 - **Alcohol**: valores omissos serão tratados após a unificação do dataset com os dados do Vivino, o qual contém também informação sobre teor alcóolico.
 - **Price**: Os valores omissos foram preenchidos com o preço médio dos vinhos da mesma vinícola (Winery). Para isso foi necessário:
   - Converter o campo **Price** de *object* para *float64*.
   - Converter o campo **Bottle Size** de *object* para *float64* e tratar os valores.
   - Recalcular o preço para que todos os registros estejam no mesmo padrão de uma garrafa de 750ml (tamanho padrão). 5,5% das análises de vinho tinto publicadas no site foram feitas com tamanhos diferentes de garrafa.
   - Calcular o preço médio por vinícola e aplicar aos registros com preço faltantes.
   - Após preencher o preço médio por vinícola nos registros de preço faltantes, o número de registros omissos passou de 4755 para 910. Esses últimos valores foram preenchidos com a média geral do preço.
 - **Variety**: como apenas um registro omisso foi encontrado e não possível identificar a variedade da uva usada para a produção do vinho ao ler o review publicado no site, o registro foi descartado.
 - **Description**: foram encontrados 43 registros com a descrição do vinho nula; o campo de descrição é parte fundamental para a construção dos modelos de aprendizagem de seção 5 e, portanto, os registros nulos foram descartados.
 - **Designation**, **Taster**: valores omissos não foram tratados, pois - por trata-se de colunas não aproveitadas para a construção dos modelos preditivos, sendo a coluna **Designation** utilizada na seção 3.4 para cruzamento dos datasets.

values_missing(winemag)

# Variety
winemag = winemag.dropna(subset=['Variety'])

# Price: O preço médio é calculado por vinícola e aplicado aos registros de preço faltante.
winemag['Price'] = winemag.groupby('Winery')['Price'].transform(lambda val: val.fillna(val.mean()))

values_missing(winemag)

# Preenche os demais preços nulos do dataset com a média do preço
winemag['Price'] = winemag['Price'].fillna(winemag['Price'].mean())

# Elimina os registros cuja coluna de descrição seja nula
winemag.dropna(subset=['Description'], inplace=True)

values_missing(winemag)

### Remoção de colunas
- As colunas **Category** e **URL** foram removidas do dataset. A categoria foi usada para filtrar o dataset para somente vinhos tintos, portanto pode ser removida pois os valores da coluna possuem o mesmo valor. A coluna URL foi gravada para facilitar identificar possíveis erros oriundos da coleta e não terá mais uso de agora em diante.

winemag = winemag.drop(['Category', 'URL'], axis=1)

winemag.info()

### Exportação do Dataset

- Dos 87.574 registros de vinhos tintos inicialmente (após a filtragem), após a tratamento dos registros duplicados e omissos têm-se 87.216 registros.

# Exporta o dataset para arquivo csv após o tratamento dado até o momento
winemag.to_csv('datasets/winemag_red.csv', index=False, header=True)
