==> 2 - Vivino - Processamento e Tratamento de Dados.ipynb <==
### Etapas iniciais: importação de pacotes, definição de funções, carregamento de dados

import pandas as pd

def values_missing(df, only_missing=True):
    '''Calcula e retorna o percentual e o número de registros nulos no dataframe df'''
    missing = pd.concat({'percent_missing': 100* df.isnull().sum() / len(df),
                         'rows_missing': df.isnull().sum()}, axis=1)
    if only_missing:
        missing = missing[missing.percent_missing > 0]
    return missing.sort_values(by='percent_missing', ascending=False)


# Carregamento dos datasets
vivino_wines = pd.read_csv('datasets/vivino_wines.csv')
vivino_vintages = pd.read_csv('datasets/vivino_vintages.csv')

## Dataset: Vivino Wines

# Dataset com os dados de vinhos
vivino_wines.info()

### Filtragem
- O dataset originalmente continha 343.176 registros, os quais cobriam diversos tipos/categorias, como vinhos brancos, espumantes e tintos. Como o trabalho restringe-se a vinhos tintos, nessa etapa foram filtrados apenas os vinhos dessa categoria resultando em 210.432 registros.

vivino_wines = vivino_wines[vivino_wines['type_id'] == 1]

vivino_wines.info()

### Tratamento de dados duplicados
- Foram encontrados e removidos 1.666 registros inteiramente duplicados.
- Após isso verificou-se a existencia de 1.370 registros duplicados pelo campo **full_name**. Ao analisar pontualmente os registros duplicados, acessando também as páginas correspondentes aos vinho no site do Vivino, pode-se perceber que que os vinhos estão duplicados no site do Vivino, sendo que em alguns casos ao acessar uma das urls é feito o redirecionamento da página para a outra, indicando possivelmente tratar-se de um registro ser o da atual e o outro um antigo. Desta forma optou-se por manter apenas o registro com maior número de análises feitas pelos usuários (**ratings_count**), por entender que o registro com maior número de avaliações é o mais atual. Os 1.370 foram removidos com a aplicação desse entendimento.

# Remoção de registros inteiramente duplicados
print('vivino_wines: Foram encontrados {} registros inteiramente duplicados'.format(vivino_wines.duplicated().sum()))

vivino_wines = vivino_wines.drop_duplicates()
print('Total de registros após a remoção dos registro duplicados: {}'.format(vivino_wines.shape[0]))

# Registros duplicados pelo **full_name**
vivino_wines[ vivino_wines.duplicated(subset=['full_name'], keep=False) ].sort_values(by='full_name').tail(50)

# Imprimindo para verificação as URLs de alguns dos vinhos duplicados pelo atributo full_name
url_ids = [237297, 247021, 69060, 327087, 224677, 245179, 172785, 54512, 223193, 21787]

with pd.option_context('display.max_colwidth', 150):
    for url in url_ids:
        print('{}\n{}'.format(vivino_wines.loc[url, ['full_name']], vivino_wines.loc[url, ['url']]))

vivino_wines[vivino_wines.duplicated(subset=['full_name'],keep=False)]\
            .sort_values(by=['full_name', 'ratings_count'],ascending=[True, False])

vivino_wines = vivino_wines.sort_values(by=['full_name', 'ratings_count'], ascending=[True, False])\
            .drop_duplicates(subset=['full_name'], keep='first')

### Tratamento de dados omissos
- Foram encontrados valores omissos em 4 colunas e dados os seguintes tratamentos:
 - **Fizziness**: a coluna em questão indica o nível de presença de gás (efervescência) do vinho espumante em uma escala de 1 (delicado) a 5 (com gás). Tal medida é própria para espumantes e isso explica o porquê da ausência de 99.99% do preenchimento do campo após a filtragem do dataset para vinhos tintos. A coluna foi removida do dataset.
 - **Tannin**: Os 41 registros omissos (0,02%) foram preenchidos com a média.
 - **Sweetness**: Os 8 registros omissos (0.004%) foram preenchidos com a média.
 - **Name**: Apenas 1 registro omisso. Ao analisar a imagem do rótulo do vinho no site do Vivino percebeu-se que o nome do vinho é 'NA' e estava sendo interpretado como registro nulo. O registro foi corrigido para o nome correto.

values_missing(vivino_wines)

# Remoção da coluna fizziness - própria de espumantes
vivino_wines = vivino_wines.drop('fizziness', axis=1)

# Tannin: registros omissos preenchidos com a média
vivino_wines['tannin'] = vivino_wines['tannin'].fillna(vivino_wines['tannin'].mean())
# Sweetness: registros omissos preenchidos com a média
vivino_wines['sweetness'] = vivino_wines['sweetness'].fillna(vivino_wines['sweetness'].mean())

# Name
vivino_wines[vivino_wines['name'].isnull()]

# URL do vinho com nome ausente
with pd.option_context('display.max_colwidth', 150):
    print(vivino_wines[vivino_wines['name'].isnull()].url)

# Corrigindo nome "ausente"
vivino_wines.loc[220170, 'name'] = 'Monte da Ravasqueira NA'
vivino_wines.loc[[220170]]

values_missing(vivino_wines)

### Remoção de colunas
- As colunas **type_id**, **type** e **url** foram removidas do dataset. As colunas de tipo foram usadas para filtrar o dataset para somente vinhos tintos, portanto pode ser removida pois os valores das colunas possuem sempreo mesmo valor. A coluna url foi gravada para facilitar identificar possíveis erros oriundos da coleta e não terá mais uso de agora em diante.

vivino_wines = vivino_wines.drop(['type_id', 'type', 'url'], axis=1)

## Dos 210.432 registros de vinhos tintos inicialmente (após a filtragem), após a tratamento dos registros duplicados 
## e omissos têm-se 207.396 registros.
vivino_wines.info()

## Dataset: Vivino Vintages

# Dataset com os dados de safras/ano de engarrafamento
vivino_vintages.info()

### Filtragem
- nenhuma filtragem foi necessária, pois após a unificação de todos os datasets do Vivino, a filtragem feita no dataset Vivino Wines garantirá que apenas vinhos tintos estejam no dataset final.

### Tratamento de dados duplicados
- Foram encontrados e removidos 9418 registros inteiramente duplicados. Reduzind os 1.290.886 registros iniciais para 1.281.468 registros.

# Remoção de registros inteiramente duplicados
print('vivino_vintages: Foram encontrados {} registros inteiramente duplicados'.format(vivino_vintages.duplicated().sum()))

vivino_vintages = vivino_vintages.drop_duplicates()
print('Total de registros após a remoção dos registro duplicados: {}'.format(vivino_vintages.shape[0]))

print('vivino_vintages: Foram encontrados {} registros duplicados pelo ID'\
      .format(vivino_vintages.duplicated(subset=['id']).sum()))

print('vivino_vintages: Foram encontrados {} registros duplicados pelo ID do vinho e ano de engarrafamento/safra'\
      .format(vivino_vintages.duplicated(subset=['wine_id', 'year']).sum()))

### Tratamento de dados omissos
- Não foram encontrados dados omissos em todo o dataset.

values_missing(vivino_vintages)

## Unificação dos datasets da Vivino
- Os 207.396 registros de vinhos tintos presentes no dataset Vivino Wines foram enriquecidos com as informações de safras específicas/ano engarramento de VIvino Vintages, o que levou o dataset a 874.676 registros. Importante notar que agora a unicidade do vinho deixa de ser apenas pelo **full_name**, sendo que a partir de agora deve levar em conta o atributo **year**. Isso permitirá na próxima a junção com o dataset da Wine Enthusiast Magazine.

# Merge do dataset Vivino Wines com Vivino Vintages. A coluna id do segundo dataset foi renomeada para vintage_id,
# pois já existe outra coluna id no primeiro dataset. Após o merge, a coluna wine_id usada para o merge foi descartada.
vivino_red = pd.merge(vivino_wines, vivino_vintages.rename({'id': 'vintage_id'}, axis=1), left_on='id', right_on='wine_id')\
            .drop('wine_id', axis=1)


vivino_red.info()

### Exportação do dataset

# Exporta o dataset para arquivo csv após o tratamento dado até o momento
vivino_red.to_csv('datasets/vivino_red.csv', index=False, header=True)
