==> 3 - Red Wines - junção dos datasets Winemag e Vivino.ipynb <==
### Etapas iniciais: importação de pacotes, definição de funções, carregamento de dados

import pandas as pd
import numpy as np

# Carregamento dos datasets
winemag = pd.read_csv('datasets/winemag_red.csv')
vivino = pd.read_csv('datasets/vivino_red.csv')

def values_missing(df, only_missing=True):
    '''Calcula e retorna o percentual e o número de registros nulos no dataframe df'''
    missing = pd.concat({'percent_missing': 100* df.isnull().sum() / len(df),
                         'rows_missing': df.isnull().sum()}, axis=1)
    if only_missing:
        missing = missing[missing.percent_missing > 0]
    return missing.sort_values(by='percent_missing', ascending=False)


## Junção dos datasets da Wine Enthusiast e do Vivino
- A junção dos dataset será feita considerando que a chave primária do dataset Vivino é composta pelos campos **full_name** e **year**. Primeiramente são feitas transformações no dataset Winemag para que possuir campos equivalentes aos campos chave do Vivino. Para permitir isso, foram feitos os seguintes passos:
 - 1° passo: Extrair da coluna **Title** do dataset Winemag e informação de ano/safra, armazenar a informação em uma  coluna nova denomiada **year**.
 - 2° passo: Analisar registros individuais de ambos os datasets procurando encontrar padrões para identificar o campo **full_name** do Dataset Vivino a partir dos dados disponíveis no dataset Winemag (já que não há um campo de identificador comum entre os datasets).
 - 3° passo: Para cada nome candidato, criar a coluna com o nome e juntar os datasets. 
 - 4° passo: Unificar todos os datasets gerados no passo 3.

#### Winemag: criação da coluna **year** a partir da extraçào de informação contida em **Title**

# Analisando os samples do título, encontra-se o padrão regex (\s\d{4}\s) para a identificação de grupos de anos
with pd.option_context('display.max_colwidth', 150):
    print(winemag.Title.sample(20))

# Dos 87.259 registros, 381 que não possuem o padrão de ano no campo Title. Após a extração dos dados 
# esses vinhos ficarão com a nova coluna year nula e não serão levados na junção com o dataset Vivino.
condition = winemag.Title.str.contains(r'(\s\d{4}\s)', regex=True)
with pd.option_context('display.max_colwidth', 150):
    print(winemag.Title[ ~condition].sample(20))

# Extração e criação da nova coluna year
winemag['year'] = winemag.Title.str.extract(r'(\s\d{4}\s)').replace(' ', '').astype('float64')

winemag.info()

winemag.loc[:, ['year', 'Title']].sample(10)

#### Winemag: criação de colunas candidatas a junção com a coluna full_name(Vivino)

- Padrões identificados na coluna full_name para criação dos nomes candidatos, com exemplos.
| # | Title (Winemag) - example | full_name (Vivino) - example | name composition - example| candidate name pattern | matches |
| --- | --- | --- | --- | --- | --- |
| 1 | Scott Paul 2016 Maresh Vineyard Pinot Noir (Dundee Hills) | Scott Paul Maresh Vineyard Pinot Noir  | *Winery*: Scott Paul <br> *Designation*: Maresh Vineyard <br> *Variety*: Pinot Noir | [Winery] + [Designation] + [Variety] | 7.042 |
| 2 | Marco Abella 2013 Mas Mallola Red (Priorat) | Marco Abella Mas Mallola | *Winery*: Marco Abella <br> *Designation*: Mas Mallola | [Winery] + [Designation] | 2.861 |
| 3 | Saddleback 2011 Cabernet Sauvignon (Oakville) | Saddleback Cabernet Sauvignon | *Winery*: Saddleback <br> *Variery*: Cabernet Sauvignon | [Winery] + [Variety] | 8.892 |
| 4 | Apaltagua 2015 Reserva Carmenère (Apalta) | Apaltagua Carmenère Reserva  | *Winery*: Apatagua <br> *Designation*: Reserva <br> *Variety*: Carmenère | [Winery] + [Variety] + [Designation] | 272 |
| 5 | Château Feytit-Clinet 2017 Barrel Sample  (Pomerol) | Château Feytit-Clinet Pomerol | *Winery*: Château Feytit-Clinet <br> *Appellation*: Pomerol, Bordeaux, France | [Winery] + primeira parte de [Appellation] | 6.256 |
| 6 | Château du Caillau 2014 Cahors | Château du Caillau Malbec Cahors | *Winery*: Château du Caillau <br> *Variety*: Malbec <br> *Appellation*: Cahors, Southwest France, France | [Winery] + [Variety] + primeira parte de [Appellation] | 176 |
| 7 | Chime 2013 Cabernet Sauvignon (California) | Chime California Cabernet Sauvignon 	| *Winery*: Chime <br> *Appellation*: California, California Other, California, US <br> *Variety*: Cabernet Sauvignon | [Winery] + primeira parte de [Appellation] + [Variety] | 876 |
| 8 | Marcarini 2013 La Serra (Barolo) | Marcarini La Serra Barolo | *Winery*: Marcarini <br> *Designation*: La Serra <br> *Appellation*: Roero, Piedmont, Italy | [Winery] + [Designation] + primeira parte de [Appellation] 	 | 1.310 |
| 9 | Demarie 2011 Riserva (Roero) | Demarie Roero Riserva | *Winery*: Demarie <br> *Appellation*: California, California Other, California, US <br> *Designation*: Riserva | [Winery] + primeira parte de [Appellation] + [Designation] | 639 |
| 10 | Algodon 2016 Estate Bonarda (Mendoza) | Algodon Estate Bonarda | *Winery*: Algodon <br> *Designation*: Estate <br> *Variety*: Bonarda, Other Red | [Winery] + [Designation] + primeira parte de [Variety] | 360 |
| 11 | Lamadrid 2015 Single Vineyard Bonarda (Agrelo) | Lamadrid Bonarda Single Vineyard | *Winery*: Lamadrid <br> *Variety*: Bonarda, Other Red <br> *Designation*: Single Vineyard | [Winery] + primeira parte de [Variety] + [Designation] | 73 |
| 12 | Planeta 2013 Frappato (Vittoria) | Planeta Frappato | *Winery*: Planeta <br> *Variety*: Frappato, Italian Red | [Winery] + primeira parte de [Variety] | 549 |
| 13 | Scala Dei 2016 Garnatxa (Priorat) | Scala Dei Garnatxa Priorat | *Winery*: Scala Dei <br> *Variety*: Garnatxa, Grenache <br> *Appellation*: Priorat, Catalonia, Spain | [Winery] + primeira parte de [Variety] + primeira parte de [Appellation] | 35 |
| 14 | Kirkland Signature 2014 Meritage (Rutherford) | Kirkland Signature Rutherford Meritage | *Winery*: Kirkland Signature <br> *Appellation*: Rutherford, Napa, California, US <br> *Variety*: Meritage, Bordeaux-style Red Blend | [Winery] + primeira parte de [Appellation] + primeira parte de [Variety] | 55 |
| **Total** | --- | --- | --- | --- | **29.396** |

# Imprimindo exemplos de títulos para investigar os padrões de nomenclatura do Vivino
with pd.option_context('display.max_colwidth', 150):
    print(winemag.Title.sample(10))

winemag.loc[32091]
#winemag[ winemag['Winery'] == '']

vivino[vivino['full_name'] == 'V. Sattui Napa Valley Syrah'].head(1)

# Removendo espaços em branco do início e do fim de cada coluna usada
winemag['Winery'] = winemag['Winery'].str.strip()
winemag['Variety'] = winemag['Variety'].str.strip()
winemag['Designation'] = winemag['Designation'].str.strip()
winemag['Appellation'] = winemag['Appellation'].str.strip()
vivino['full_name'] = vivino['full_name'].str.strip()

# Criação de novas colunas com apenas a primeira parte da informação das colunas Appellation e Variety
winemag['first_appellation'] = winemag['Appellation'].str.extract(r'([\w ]+),', expand=False)
winemag['first_variety'] = winemag['Variety'].str.extract(r'([\w ]+),', expand=False)

# Definição da lista com os possíveis padrões para identificação do nome dos vinho no dataset da Vivino 
# a partir dos dados da Wine Enthusiast
candidate_patterns = [
    ['Winery', 'Designation', 'Variety'],
    ['Winery', 'Designation', ''],
    ['Winery', 'Variety', ''],
    ['Winery', 'Variety', 'Designation'],
    ['Winery', 'first_appellation', ''],
    ['Winery', 'Variety', 'first_appellation'],
    ['Winery', 'first_appellation', 'Variety'],
    ['Winery', 'Designation', 'first_appellation'],
    ['Winery', 'first_appellation', 'Designation'],
    ['Winery', 'Designation', 'first_variety'],
    ['Winery', 'first_variety', 'Designation'],
    ['Winery', 'first_variety', ''],
    ['Winery', 'first_variety', 'first_appellation'],
    ['Winery', 'first_appellation', 'first_variety']
]
red_wines_dfs = []
i = 0
for cp in candidate_patterns:
    if cp[2]:
        winemag['candidate'] = winemag.apply(\
                     lambda row: '{} {} {}'.format(row[cp[0]], row[cp[1]], row[cp[2]]) , axis = 1)
    else:
        winemag['candidate'] = winemag.apply(\
                     lambda row: '{} {}'.format(row[cp[0]], row[cp[1]]) , axis = 1)
    red_wines_dfs.append(\
        pd.merge(winemag, vivino,  how='inner', left_on=['candidate','year'], right_on = ['full_name','year']))
    # Atualiza dataset Winemag removendo os registro já unificados com Vivino
    winemag = winemag[~winemag['Title'].isin(list(red_wines_dfs[i].Title))]
    i += 1

i = 0
for df in red_wines_dfs:
    print('Pattern : {}\nEncontrados e combinados {} registros.\n'.format(candidate_patterns[i], red_wines_dfs[i].shape[0]))
    i += 1

# Unificação dos datasets gerados a partir dos padrões identificados
red_wines = pd.concat(red_wines_dfs).drop(['candidate', 'first_appellation', 'first_variety'],axis=1)\
                .reset_index() 

print('Número de registros combinados: {}'.format(red_wines.shape[0]))
print('Percentual de registros combinados: {0:.2%}'.format((red_wines.shape[0]) / 87259))    

### Tratamento de dados omissos
- O campo **Alcohol** (Winemag) possui 2.31% dos registros (680) com valor nulo. O campo **alcohol** (Vivino), a uma primeira vista, não possui valores nulos, porém 8.437 registros possuem valor zero; e na verdade são oriundos de valores omissos.
 - O campo **alcohol** foi usado para preencher os valores omissos de **Alcohol**. Os valores preenchidos com dados inconsistentes (maiores que 23% ou menos que 5.5%) foram substituídos pela média do dataset. 
   - [5.5% - 23%] - Intervalo de variação para teor alcoólico de vinhos. https://winefolly.com/tips/the-lightest-to-the-strongest-wine/
- Os campos **Taster** e **Designation** possuem alguns valores nulos, mas não foram tratados pois não serão usados para a construção dos modelos de aprendizagem.

# Verificando registros duplicados
print('Duplicados em todas as colunas: ', red_wines.duplicated().sum())
print('Duplicados pelo título: ', red_wines.duplicated(subset=['Title']).sum())
print('Duplicados pelas colunas usadas para identificação: ', red_wines.duplicated(subset=['year', 'Winery', 'Appellation', 'Variety', 'Designation']).sum())

# Verificando registros omissos
values_missing(red_wines, only_missing=False)

(red_wines['alcohol'] == 0).sum()

# Preenche valores omissos de Alcohol com os dados de alcohol 
# e substituí valores inconsistentes (menores que 5.5% ou maiores que 23%) pela média
alcohol = red_wines.apply(lambda x: x['alcohol'] if pd.isnull(x['Alcohol']) else x['Alcohol'], axis=1).\
    apply(lambda x: x if (x >= 5.5) and (x <= 23) else np.nan)
red_wines['Alcohol'] = alcohol.fillna(alcohol.mean())

red_wines['Alcohol'].describe()

### Correções e transformações nos dados
 - Criação da coluna **continent** aplicando a função *country_to_continent*. Sintetizei a informação do páis, criando a coluna com informação do continente, com o intuito de ter dados mais generalizados para uso nos modelos. Percebeu-se ao analisar a média por país e a média continental da avaliação do vinho que os valores por país em cada continente são relativamente semelhantes.

print(red_wines.country.value_counts())
print('\n', red_wines.groupby('country')['Rating'].mean())

def country_to_continent(country):
    '''Retorna o continente a partir do nome do país'''
    south_america = ['Argentina', 'Brazil', 'Chile', 'Uruguay']
    north_america = ['Canada', 'United States']
    oceania = ['Australia', 'New Zealand']
    africa = ['South Africa']
    asia = ['Israel']
    if country in south_america:
        return 'South America'
    elif country in north_america:
        return 'North America'
    elif country in oceania:
        return 'Oceania'
    elif country in africa:
        return 'Africa'
    elif country in asia:
        return 'Asia'
    else:
        return 'Europe'

red_wines['continent'] = red_wines['country'].apply(country_to_continent)

red_wines.groupby('continent')['Rating'].mean()

- Criacão das colunas **popular_wine** e **popular_vintage** a partir dos campos **ratings_count** e **vinage_ratings_count**, a fim de representar se determinado vinho e safra, respectivamente, são populares (popular definido como sendo número de avaliações maior do que a média).

red_wines['popular_wine'] = red_wines['ratings_count'].apply(\
                                lambda x: 1 if (x >= red_wines['ratings_count'].mean()) else 0)
red_wines['popular_vintage'] = red_wines['vintage_ratings_count'].apply(\
                                lambda x: 1 if (x >= red_wines['vintage_ratings_count'].mean()) else 0)

- Criação da coluna **rating_group** com base no valor de Rating agrupado de acordo com o guia de interpretação da nota publicado no site da Wine Enthusiast. <br><br>
 
- Grupo **0** --> 98–100: Classic "The pinnacle of quality"
- Grupo **1** --> 94–97: Superb "A great achievement" 
- Grupo **2** --> 90–93: Excellent "Highly recommended"
- Grupo **3** --> 87–89: Very Good "Often good value; well recommended"
- Grupo **4** --> 83–86: Good "Suitable for everyday consumption; often good value"
- Grupo **5** --> 80–82: Acceptable "Can be employed in casual, less-critical circumstances"
<br>
https://www.winemag.com/our-buying-guide-and-blind-tasting-process/

# Conversão de ratings para rating group

def ratings_to_group(rating):
    if rating in range(80,83):
        return 0
    elif rating in range(83,87):
        return 1
    elif rating in range(87,90):
        return 2
    elif rating in range(90,94):
        return 3
    elif rating in range(94,98):
        return 4
    else:
        return 5
    
red_wines['rating_group'] = red_wines['Rating'].apply(ratings_to_group)

### Remoção de colunas
- Os seguintes campos foram removidas do dataset pelos motivos abaixo:
 - Campos de identificação do vinho/safra: **index**, **id**, **vintage_id** e **Title**.
 - Campos com informações textuais do vinho e do review que não serão transformadas em categorias: **Taster**, **name**, **wine_region**, **winery** e **winery_region**.
 - Campos usados para unificação dos datasets: **Designation**, **Variety**, **Appellation**, **Winery**, **full_name** e **year**.
 - Campos usados em transformações de outros valores: **Bottle Size**, **alcohol**, **ratings_count**, **vintage_ratings_count**, **country** e **Rating**.

red_wines.drop(
    ['index', 'id', 'vintage_id', 'Title',
     'Taster', 'name','wine_region', 'winery', 'winery_region',
     'Designation', 'Variety', 'Appellation', 'Winery', 'full_name', 'year',
     'Bottle Size', 'alcohol', 'country', 'ratings_count','vintage_ratings_count', 'Rating'], 
    axis = 1, inplace=True
)

### Exportação do dataset

# Exporta o dataset para arquivo csv após o tratamento dado
#red_wines.to_csv('datasets/red_wines.csv', index=False, header=True)

red_wines.info()
